// src/types/tag.ts
enum Tag {
    TypeScript = "typescript",
    JavaScript = "javascript",
    Git = "git",
    GitHub = "github",
    React = "react",
    SQL = "sql",
    Eleventy = "eleventy",
    Python = "python",
    HTML = "html",
    CSS = "CSS",
    Express = "express",
    NodeJS = "nodejs",
    Bun = "bun",
    Hono = "hono",
    MongoDB = "mongodb",
    Jest = "jest",
    Vitest = "vitest",
    Vite = "vite",
    Figma = "figma",
    Jira = "jira",
    Vue = "vue",
    Redux = "redux",
    Frontend = "frontend",
    Backend = "backend",
    Database = "database",
    Testing = "testing",
    Deployment = "deployment",
    Collaboration = "collaboration",
    API = "api",
    Bootstrap = "bootstrap",
    TailwindCSS = "tailwindcss",
    NextJS = "nextjs",
    PostgreSQL = "postgresql",
    OAuth = "oauth"
};

export default Tag;